requirements.txt
boto3==1.28.11
python-dotenv==1.0.0
ipython==8.14.0